pyproject.toml
[project]
name = "cv-tracker"
version = "0.1.0"
description = "CV Tracker & Smart ATS Matcher"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "gunicorn>=22.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # File Parsing
    "PyPDF2>=3.0",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    # LLM
    "groq>=0.4",
    # Filesystem Watching
    "watchdog>=4.0",
    # Export
    "openpyxl>=3.1",
    "reportlab>=4.1",
    # Frontend
    "streamlit>=1.35",
    "httpx>=0.27",
    # File upload
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["backend*", "frontend*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
